requirements.txt
ezdxf>=1.1.0
pyproj>=3.4.0
pandas>=1.5.0
pyyaml>=6.0
numpy>=1.23.0
